requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
aiofiles==23.2.1
asyncio-mqtt==0.16.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0 